# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Inventory API Documentation",
        "description": "API documentation for the Inventory project",
        "contact": {
            "email": "your-email@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Inventory API Server"
        }
    ],
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Connexion utilisateur",
                "description": "Retourne l'utilisateur et un token Sanctum",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "login",
                                    "password"
                                ],
                                "properties": {
                                    "login": {
                                        "type": "string",
                                        "example": "user@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Identifiants invalides"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Créer un utilisateur",
                "description": "Crée un utilisateur lié à une entreprise",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "phone_number",
                                    "password",
                                    "password_confirmation",
                                    "id_entreprise"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "phone_number": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "type": "string"
                                    },
                                    "id_entreprise": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "profil_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "type": "string"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Déconnexion",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "Déconnexion réussie"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Entreprise": {
                "title": "Entreprise",
                "description": "Informations sur une entreprise",
                "properties": {
                    "id_entreprise": {
                        "description": "Identifiant unique de l'entreprise",
                        "type": "string",
                        "format": "uuid"
                    },
                    "company_name": {
                        "description": "Nom de l'entreprise",
                        "type": "string",
                        "maxLength": 255
                    },
                    "tax_number": {
                        "description": "Numéro fiscal",
                        "type": "string",
                        "nullable": true
                    },
                    "address": {
                        "description": "Adresse de l'entreprise",
                        "type": "string",
                        "nullable": true
                    },
                    "company_email": {
                        "description": "Email de l'entreprise",
                        "type": "string",
                        "format": "email",
                        "nullable": true
                    },
                    "logo": {
                        "description": "Logo (URL ou chemin)",
                        "type": "string",
                        "nullable": true
                    },
                    "manager_name": {
                        "description": "Nom du responsable",
                        "type": "string",
                        "nullable": true
                    },
                    "manager_email": {
                        "description": "Email du responsable",
                        "type": "string",
                        "format": "email",
                        "nullable": true
                    },
                    "manager_phone": {
                        "description": "Téléphone du responsable",
                        "type": "string",
                        "nullable": true
                    },
                    "manager_nui": {
                        "description": "NUI du responsable",
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Date de création",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date de mise à jour",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Profil": {
                "title": "Profil",
                "properties": {
                    "id_profil": {
                        "description": "Profile's unique identifier",
                        "type": "string",
                        "format": "uuid"
                    },
                    "nom": {
                        "description": "Profile's name",
                        "type": "string"
                    },
                    "description": {
                        "description": "Profile's description",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "title": "User",
                "description": "Utilisateur du système",
                "properties": {
                    "id_users": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "id_entreprise": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "phone_number": {
                        "type": "string"
                    },
                    "agence": {
                        "type": "string",
                        "nullable": true
                    },
                    "poste": {
                        "type": "string",
                        "nullable": true
                    },
                    "link_img": {
                        "type": "string",
                        "nullable": true
                    },
                    "matricule": {
                        "type": "string",
                        "nullable": true
                    },
                    "signature": {
                        "type": "string",
                        "nullable": true
                    },
                    "entreprise": {
                        "$ref": "#/components/schemas/Entreprise"
                    },
                    "profil": {
                        "$ref": "#/components/schemas/Profil"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}